+++ config/i18n.json
@@ -25,6 +25,17 @@
     "Normal": "Normal",
     "WeeklyHeroic": "Weekly Heroic Strike"
   },
+  "Armory": {
+    "WishlistedRolls": "Wishlisted Roll",
+    "WishlistedRolls_plural": "{{count, number}} Wishlisted Rolls",
+    "TrashlistedRolls": "Trashlisted Roll",
+    "TrashlistedRolls_plural": "{{count, number}} Trashlisted Rolls",
+    "NoNotes": "No Notes",
+    "NoItems": "You don't have any of these items",
+    "YourItems": "Your Items",
+    "Unknown": "Unknown Item",
+    "Season": "Season {{season}}, Year {{year}}"
+  },
   "Bucket": {
     "Armor": "Armor",
     "General": "General",
