+++ src/app/armory/AllWishlistRolls.m.scss
@@ -1,0 +1,4 @@
+.roll {
+  display: flex;
+  flex-direction: row;
+}
