+++ src/app/armory/AllWishlistRolls.m.scss.d.ts
@@ -1,12 +1,7 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'armory': string;
-  'flavor': string;
-  'header': string;
-  'section': string;
-  'source': string;
-  'subtitle': string;
+  'roll': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
