+++ src/app/armory/Armory.tsx
@@ -1,14 +1,11 @@
 import { DestinyAccount } from 'app/accounts/destiny-account';
 import { addCompareItem } from 'app/compare/actions';
-import { languageSelector } from 'app/dim-api/selectors';
 import BungieImage, { bungieNetPath } from 'app/dim-ui/BungieImage';
 import ElementIcon from 'app/dim-ui/ElementIcon';
-import ExternalLink from 'app/dim-ui/ExternalLink';
 import RichDestinyText from 'app/dim-ui/RichDestinyText';
 import ShowPageLoading from 'app/dim-ui/ShowPageLoading';
 import { t } from 'app/i18next-t';
 import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
-import { DimItem } from 'app/inventory/item-types';
 import ItemIcon from 'app/inventory/ItemIcon';
 import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
 import { allItemsSelector, bucketsSelector, storesLoadedSelector } from 'app/inventory/selectors';
@@ -17,35 +14,24 @@ import { useLoadStores } from 'app/inventory/store/hooks';
 import { getEvent, getSeason } from 'app/inventory/store/season';
 import EmblemPreview from 'app/item-popup/EmblemPreview';
 import { hideItemPopup } from 'app/item-popup/item-popup';
-import { LoreLink } from 'app/item-popup/ItemDescription';
-import { AmmoIcon, destinyDBLink, ItemTypeName } from 'app/item-popup/ItemPopupHeader';
+import { AmmoIcon, ItemTypeName } from 'app/item-popup/ItemPopupHeader';
 import ItemSockets from 'app/item-popup/ItemSockets';
 import ItemStats from 'app/item-popup/ItemStats';
 import MetricCategories from 'app/item-popup/MetricCategories';
-import Plug from 'app/item-popup/Plug';
 import { useDefinitions } from 'app/manifest/selectors';
 import Objective from 'app/progress/Objective';
 import { Reward } from 'app/progress/Reward';
 import { AppIcon, compareIcon, faClock } from 'app/shell/icons';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
 import { getItemYear } from 'app/utils/item-utils';
-import { wishListRollsForItemHashSelector } from 'app/wishlists/selectors';
-import { WishListRoll } from 'app/wishlists/types';
 import clsx from 'clsx';
 import { D2EventInfo } from 'data/d2/d2-event-info';
 import { ItemCategoryHashes } from 'data/d2/generated-enums';
-import _ from 'lodash';
 import React from 'react';
 import { useSelector } from 'react-redux';
+import AllWishlistRolls from './AllWishlistRolls';
 import styles from './Armory.m.scss';
-
-const links = {
-  'Light.gg': (item: DimItem, language: string) =>
-    `https://www.light.gg/db/${language}/items/${item.hash}`,
-  DestinyTracker: destinyDBLink,
-  'data.destinysets.com': (item: DimItem, language: string) =>
-    `https://data.destinysets.com/i/InventoryItem:${item.hash}?lang=${language}`,
-};
+import Links from './Links';
 
 export default function Armory({
   account,
@@ -60,8 +46,6 @@ export default function Armory({
   useLoadStores(account, storesLoaded);
   const buckets = useSelector(bucketsSelector)!;
   const allItems = useSelector(allItemsSelector);
-  const wishlistRolls = useSelector(wishListRollsForItemHashSelector(itemHash));
-  const language = useSelector(languageSelector);
 
   if (!storesLoaded || !defs) {
     return <ShowPageLoading message={t('Loading.Profile')} />;
@@ -78,7 +62,7 @@ export default function Armory({
   if (!item) {
     return (
       <div className="dim-page">
-        <h1>Unknown Item</h1>
+        <h1>{t('Armory.Unknown')}</h1>
       </div>
     );
   }
@@ -96,9 +80,7 @@ export default function Armory({
   const screenshot = itemDef.screenshot;
   const flavorText = itemDef.flavorText || itemDef.displaySource;
 
-  console.log({ item });
-
-  const [goodRolls, badRolls] = _.partition(wishlistRolls, (r) => !r.isUndesirable);
+  // TODO: Show Catalyst benefits for exotics
 
   return (
     <div
@@ -113,18 +95,7 @@ export default function Armory({
           : undefined
       }
     >
-      <ul className={styles.links}>
-        {_.map(links, (link, name) => (
-          <li key={name}>
-            <ExternalLink href={link(item, language)}>{name}</ExternalLink>
-          </li>
-        ))}
-        {item.loreHash && (
-          <li>
-            <LoreLink loreHash={item.loreHash} />
-          </li>
-        )}
-      </ul>
+      <Links item={item} />
       <div className={styles.header}>
         <div className="item">
           <ItemIcon item={item} />
@@ -149,8 +120,9 @@ export default function Armory({
                 {season.displayProperties.hasIcon && (
                   <BungieImage height={15} width={15} src={season.displayProperties.icon} />
                 )}{' '}
-                {season.displayProperties.name} (Season {season.seasonNumber}, Year{' '}
-                {getItemYear(item)}){event && <span> - {D2EventInfo[getEvent(item)].name}</span>}
+                {season.displayProperties.name} (
+                {t('Armory.Season', { season: season.seasonNumber, year: getItemYear(item) })})
+                {event && <span> - {D2EventInfo[getEvent(item)].name}</span>}
               </div>
             )}
           </div>
@@ -191,27 +163,27 @@ export default function Armory({
       )}
 
       {item.stats && !item.bucket.inArmor && (
-        <div className={styles.stats}>
+        <div className={styles.section}>
           <ItemStats item={item} />
         </div>
       )}
 
       {item.sockets && (
-        <div className={styles.stats}>
-          <ItemSockets item={item} />
+        <div className={styles.section}>
+          <ItemSockets item={item} minimal />
         </div>
       )}
       {item.pursuit && (
         <>
           {defs && item.objectives && (
-            <div className="item-details">
+            <div className={styles.section}>
               {item.objectives.map((objective) => (
                 <Objective objective={objective} key={objective.objectiveHash} />
               ))}
             </div>
           )}
           {defs.isDestiny2() && item.pursuit.rewards.length !== 0 && (
-            <div className="item-details">
+            <div className={styles.section}>
               <div>{t('MovePopup.Rewards')}</div>
               {item.pursuit.rewards.map((reward) => (
                 <Reward key={reward.itemHash} reward={reward} />
@@ -223,26 +195,29 @@ export default function Armory({
               <RichDestinyText text={item.pursuit.questLineDescription} ownerId={item.owner} />
             </p>
           )}
-          {itemDef.setData?.itemList.map((h) => {
-            const stepItem = makeFakeItem(defs, buckets, undefined, h.itemHash);
-            return (
-              stepItem && (
-                <div
-                  key={h.itemHash}
-                  style={{ fontWeight: h.itemHash === itemHash ? 'bold' : 'normal' }}
-                >
-                  {stepItem.name}: <RichDestinyText text={stepItem.description} />
-                </div>
-              )
-            );
-          })}
+          {itemDef.setData?.itemList && (
+            <ol>
+              {itemDef.setData.itemList.map((h) => {
+                const stepItem = makeFakeItem(defs, buckets, undefined, h.itemHash);
+                return (
+                  stepItem && (
+                    <li
+                      key={h.itemHash}
+                      style={{ fontWeight: h.itemHash === itemHash ? 'bold' : 'normal' }}
+                    >
+                      <RichDestinyText text={stepItem.description} />
+                    </li>
+                  )
+                );
+              })}
+            </ol>
+          )}
         </>
       )}
-      {item.isExotic && item.bucket.inWeapons && <h2>Catalyst??</h2>}
       {storeItems.length > 0 && (
         <>
           <h2>
-            Your Items
+            {t('Armory.YourItems')}
             {storeItems[0].comparable && (
               <button
                 className="dim-button"
@@ -252,7 +227,7 @@ export default function Armory({
                   dispatch(addCompareItem(storeItems[0]));
                 }}
               >
-                <AppIcon icon={compareIcon} /> Compare
+                <AppIcon icon={compareIcon} /> {t('Compare.Button')}
               </button>
             )}
           </h2>
@@ -266,68 +241,12 @@ export default function Armory({
                 </ItemPopupTrigger>
               ))
             ) : (
-              <div>You don't have any of these</div>
+              <p>{t('Armory.NoItems')}</p>
             )}
           </div>
         </>
       )}
-      <h2>TODO: which rolls do you have?</h2>
-      {goodRolls.length > 0 && (
-        <>
-          <h2>
-            {goodRolls.length}/
-            {item
-              .sockets!.allSockets.filter((s) => s.isPerk)
-              .reduce((combos, s) => combos * s.plugOptions.length, 1)
-              .toLocaleString()}{' '}
-            Wishlisted Rolls
-          </h2>
-          <WishlistRolls item={item} wishlistRolls={goodRolls} />
-        </>
-      )}
-      {badRolls.length > 0 && (
-        <>
-          <h2>{badRolls.length} Trashlisted Rolls</h2>
-          <WishlistRolls item={item} wishlistRolls={badRolls} />
-        </>
-      )}
+      <AllWishlistRolls item={item} />
     </div>
-  );
-}
-
-function WishlistRolls({ wishlistRolls, item }: { wishlistRolls: WishListRoll[]; item: DimItem }) {
-  const groupedWishlistRolls = _.groupBy(wishlistRolls, (r) => r.notes);
-
-  // TODO: group by making a tree of least cardinality -> most?
-
-  return (
-    <>
-      {_.map(groupedWishlistRolls, (rolls, notes) => (
-        <div key={notes}>
-          <div>{notes || 'No Notes'}</div>
-          <ul>
-            {rolls.map((r, i) => (
-              <li key={i} className={styles.roll}>
-                {Array.from(r.recommendedPerks, (h) => {
-                  const socket = item.sockets!.allSockets.find((s) =>
-                    s.plugOptions.some((p) => p.plugDef.hash === h)
-                  )!;
-                  const plug = socket.plugOptions.find((p) => p.plugDef.hash === h)!;
-                  return (
-                    <Plug
-                      key={plug.plugDef.hash}
-                      plug={plug}
-                      item={item}
-                      socketInfo={socket}
-                      hasMenu={true}
-                    />
-                  );
-                })}
-              </li>
-            ))}
-          </ul>
-        </div>
-      ))}
-    </>
   );
 }
