+++ src/app/armory/Links.m.scss
@@ -1,0 +1,17 @@
+@import '../variables';
+
+.links {
+  float: right;
+  margin: 0 0 0 32px;
+  padding: 0;
+  list-style: none;
+  @include vertical-space-children(4px);
+
+  li {
+    white-space: nowrap;
+    img {
+      vertical-align: text-bottom;
+      margin-right: 0.2em;
+    }
+  }
+}
